Implementando ControladorNegocio para a hierarquia de Controladores e revisando testes CRUD do Usuario, RelatorioFrequencia, Semana, Atividade

+++ src/br/com/projetoperiodo/model/relatorio/atividade/dao/AtividadeDao.java
@@ -15,5 +15,5 @@ public interface AtividadeDao {
 
 	public List<Atividade> listar();
 
-	public Atividade buscar(int primaryKey);
+	public Atividade buscar(long primaryKey);
 }
